15-Kubernetes/llm-demo-k8s/requirements.txt
torch
transformers
peft
datasets
accelerate
bitsandbytes
wandb 
tqdm
numpy
pandas
scikit-learn
fastapi
uvicorn
pydantic
python-dotenv